move test to NamedTopologyIntegrationTest
